+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0.9](https://github.com/smsapi/smsapi-javascript-client/compare/v2.0.8...v2.0.9) - 2023-09-14
+
+### Fixed
+
+- Build
+
 ## [2.0.8](https://github.com/smsapi/smsapi-javascript-client/compare/v2.0.7...v2.0.8) - 2023-09-12
 
 ### Fixed